pyproject.toml
[project]
name = "crm-digital-fte"
version = "0.1.0"
description = "24/7 AI Customer Success agent — Gmail, WhatsApp, Web Form support backed by PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "openai>=1.66.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["agent*", "database*"]
